refactor: rename table components
