+++ app/shared/components/tables/CompositionTable/MusicTableSection.tsx
