+++ app/shared/components/tables/WorksTable/WorkTableSection.test.tsx
@@ -1,7 +1,7 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 
-import { WorkTableSection } from './WorkTableSelection';
+import { WorkTableSection } from './WorkTableSection';
 
 import { useFetchStaticFilters } from '~/shared/hooks/use-fetch-static-filters/useFetchStaticFilters';
