+++ app/shared/components/tables/WorksTable/WorkTableSection.tsx
